test(configuration): refactor tests for env- and file-based config to separate files
ConfigFactory.test.ts was >2500 lines, which was making maint a little labourious.
No functional change.
